Do not autoload after initialization.
Rails 5 by default disables autoloading after initialization in
production. There is one spec in the fortitude test suite that depends
on this autoloading behavior, but it tests a strange use case and we are
not supporting it.

The use case is the ability to reference a constant like
Views::SomeNamespace that corresponds to an empty directory
app/views/some_namespace/ . All actual widgets are eager loaded.

Autoloading after initialization is not thread safe. For more info on
why it was disabled, see https://github.com/rails/rails/issues/13142